requirements.txt
pandas==1.0.5
pdfreader==0.1.4
pyyaml==5.3.1
psycopg2==2.8.5
boto3==1.13.25
elasticsearch==7.7.1
nltk==3.5
bs4==0.0.1
triage==4.2.0
gensim==3.8.3
sklearn==0.0
ohio==0.5.0
SQLAlchemy==1.3.18
wordcloud==1.7.0
openpyxl==3.0.6
